refactor: 코드 컨벤션 및 파일 위치 수정
