+++ CleanArea/CleanArea/iOS/Enum/StartViewType.swift
