+++ CleanArea/CleanArea/iOS/Enum/TabType.swift
