+++ CleanArea/CleanArea/iOS/Enum/TextFieldType.swift
