+++ CleanArea/CleanArea/iOS/ViewModel/Policy.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-
+/*
 class Policy: ObservableObject {
     @Published var job: [YouthPolicy] = []
     @Published var residence: [YouthPolicy] = []
@@ -27,3 +27,4 @@ class Policy: ObservableObject {
         return policies.filter { $0.polyRlmCd == code }
     }
 }
+*/
